refactor show into viewutils
